Initial cut of the PiGlow driver
This is the initial cut of the PiGlow driver. This driver is not
yet in the correct form for a GoBot metal driver. The code will need
to be refactored for that.

This code is functionally complete in terms of the driver, so should
make a good basis for the final GoBot driver.

The driver currently supports:

 *    light a single arm/spiral, so that's all 6 leds in a arm
 *    light a single ring, so that's the same colour lit in each of the
 *    three spiral arms
 *    light any random pattern of leds
 *    fade up-down or down-up over a period any single led, spiral arm
 *    or ring
 *    throb so, that's fade up-down-up or down-up-down, over a period
 *    for any led, spiral arm or ring
 *    flash n times over a period for any led, spiral arm or ring
 *    Fade, throb, or flash any random pattern of leds
 *    Turn all of the leds off

Although the driver works, is race free and includes the simulator a full
set of go test type tests still need to be created.