The simulator now quits on CTRL-C
The simulator now quits if CTRL-C or Enter or Escape is pressed
while it is running.

The required tcell based event polling code has been added to mainsim.go
and the call to time.Sleep removed from main.go.

Note: This now means that the sim runs the demo in a separate go
routine.