Remove commented out debug I2C write code
Removed the code that was in place prior to a fix to
GoBot issue #372. See:
https://github.com/hybridgroup/gobot/issues/372

The removed code wrote directly to the I2C bus using the connections
Write methods directly for both single byte values and arrays of bytes.

This was necessary until the higher level WriteBlockData and
WriteByteData were corrected. See the above bug for the details.

Now that the bug has been corrected as of GoBot v1.3.0 the code can
safely be removed.