Fix: Remove the race condition from the simulator
The channel based simulator contains a data race. The race is between
one go routine that is writing to the i2cBus channel and a second one
that is calling close on the same channel.

The fix is to ensure that only the go routine that is writing to the
channel is the one the closes the channel.

Any outstanding receives on this now closed channel will complete
immediately. This can be detected using the two value form of a channel
receive operation.

This fix has had the side effect of significantly simplifying the code
by removing the need for the done channel. Everything is now controlled
by the quit channel that is closed by a key press.